+++ pkg/providers/iam/provider.go
@@ -97,6 +97,7 @@ func (p *Subprovider) FrameworkResources() []func() resource.Resource {
 // FrameworkDataSources returns the IAM data sources implemented using terraform-plugin-framework
 func (p *Subprovider) FrameworkDataSources() []func() datasource.DataSource {
 	return []func() datasource.DataSource{
+		NewAccessibleGroupsDataSource,
 		NewAccountSwitchKeysDataSource,
 		NewAllowedAPIsDataSource,
 		NewAuthorizedUsersDataSource,