+++ pkg/providers/iam/testdata/TestDataAccessibleGroups/basic.tf
@@ -1,0 +1,7 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_iam_accessible_groups" "groups" {
+  username = "user1"
+}